requirements.txt
antlr4-python3-runtime==4.9.3
av==13.1.0
blinker==1.8.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
eva-decord==0.6.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
fsspec==2024.10.0
graphql-core==3.2.5
gunicorn==23.0.0
hydra-core==1.3.2
imagesize==1.4.1
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.2
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==11.0.0
portalocker==2.10.1
pycocotools==2.0.8
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
-e git+https://github.com/facebookresearch/sam2.git@c2ec8e14a185632b0a5d8b161928ceb50197eddc#egg=SAM_2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
six==1.16.0
strawberry-graphql==0.247.0
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
Werkzeug==3.0.6
